refactor: move intake alert to separate component
